Settings Slider updates
Added periodic update calls.
Fixed a bug that was holding slider cells in editing mode. Editing is now stopped on mouse up. This ensures that values updated manually will show slider position updates immediately instead of having to wait for a new slider in the table to be selected to stop a previous edit.