Increase size value for Horizon widgets
The layout manager isn't playing nice with the reduced size for horizon widgets. Proper resizing will occur on the vertical but not the horizontal. I've increased the preferred size value to match the current horizontal. this solves the image clipping, but the horizon widgets will be larger than other widgets for the time being.