Reworked serial sending framework
-Created a new SerialSendManager to handle outgoing serial communications, packet acknowledgement, and queuing behaviors. This is to address the potential for lost packets and missing waypoints sent to a unit, as well as to improve general handling of serial traffic.

All outgoing messages are now processed one at a time. If a message needs to be acknowledged, it will be held at the front of the sending queue until that acknowledgement comes, or the relevant timeout expires.

Once testing and any fixes reveal this change to be stable. The old SerialSender will be removed form the code base.