Feature serial management queue (#28)
* Reworked serial sending framework

-Created a new SerialSendManager to handle outgoing serial communications, packet acknowledgement, and queuing behaviors. This is to address the potential for lost packets and missing waypoints sent to a unit, as well as to improve general handling of serial traffic.

All outgoing messages are now processed one at a time. If a message needs to be acknowledged, it will be held at the front of the sending queue until that acknowledgement comes, or the relevant timeout expires.

Once testing and any fixes reveal this change to be stable. The old SerialSender will be removed form the code base.

* Address Serial edge cases

-Simplify Sync request responses and removed context.onConnection() since it was just a sendWaypointList() call on its own. This is now called directly instead of the superfluous function.
-Both the message and confirmation queue are now cleared when a serial connection is ended. This should prevent any holdover messages from being processed on a restart since the queues are initialized when the SerialSendManager is created in its constructor and not in the timers queue servicing task thread.

* Increase serial service interrupt timer

Many sends where failing out on the UI side due to inadequate length of timing. After some testing for best results, I've increased the time from 50ms to 400ms. Most messages will not take near this long, but it catches virtually all outliers.

* Fine tune serial timing

Reduced serial queue service timer and increased timeout checks for individual messages. The results are consistent with the last change, with the added benefit that the queue will empty faster if confirms are received before the 400ms timeout

* Reduce resend logging level